backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.21
pydantic>=2.0
python-dotenv==1.0.0
pydantic-settings==2.12.0
passlib
python-jose[cryptography]==3.3.0
